+++ documentation/CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changes
 
-## 16.6.0 (unreleased)
+## 17.0.0 (unreleased)
 
+* Added support for reading Akai MPC60 programs.
 * Added support for reading Akai MPC500/MPC1000/MPC2500 programs.
 * Added support for reading Akai MPC2000/MPC2000XL/MPC3000 programs.
 * Added support for reading Akai S900/S950 programs.
@@ -10,6 +11,8 @@
 * ISO File
   * New: Added support for MPC2000 format.
   * New: Shows an info text if it is a plain ISO 9660 file which can be accessed with OS functionality.  
+* Korg KMP
+  * Fixed: Velocity layers need to be stored in separate KMP files.
 * Yamaha YSFC
   * Fixed: Libraries are now limited to a max. of 128 performances.
   * Fixed: The performance names are now limited to 20 characters.